+++ common/meta/raft-store/Cargo.toml
@@ -20,11 +20,15 @@ common-meta-sled-store = { path = "../sled-store" }
 common-meta-types = { path = "../types" }
 common-tracing = { path = "../../tracing" }
 
-anyhow = "1.0.51"
+# github deps
 async-raft = { git = "https://github.com/datafuse-extras/async-raft", tag = "v0.6.2-alpha.14.1" }
+
+# crates.io deps
+anyhow = "1.0.51"
 async-trait = "0.1.52"
 clap = { version = "3.0.0-rc.7", features = ["derive", "env"] }
 bytes = "1.1.0"
+lazy_static = "1.4.0"
 derive_more = "0.99.17"
 maplit = "1.0.2"
 rand = "0.8.4"
@@ -36,6 +40,7 @@ tonic = { version = "0.6.2", features = ["transport", "codegen", "prost", "tls-r
 
 [build-dependencies]
 tonic-build = "0.6.2"
+common-building = { path = "../../building" }
 
 [dev-dependencies]
 common-base = { path = "../../base" }
