+++ common/meta/raft-store/src/config.rs
@@ -15,8 +15,34 @@
 use clap::Parser;
 use common_exception::ErrorCode;
 use common_meta_types::NodeId;
+use lazy_static::lazy_static;
 use serde::Deserialize;
 use serde::Serialize;
+
+lazy_static! {
+    pub static ref DATABEND_COMMIT_VERSION: String = {
+        let build_semver = option_env!("VERGEN_BUILD_SEMVER");
+        let git_sha = option_env!("VERGEN_GIT_SHA_SHORT");
+        let rustc_semver = option_env!("VERGEN_RUSTC_SEMVER");
+        let timestamp = option_env!("VERGEN_BUILD_TIMESTAMP");
+
+        let b = build_semver.unwrap();
+        let g = git_sha.unwrap();
+        let r = rustc_semver.unwrap();
+        let t = timestamp.unwrap();
+        println!("{}-{}-{}-{}", b, g, r, t);
+        let ver = match (build_semver, git_sha, rustc_semver, timestamp) {
+            #[cfg(not(feature = "simd"))]
+            (Some(v1), Some(v2), Some(v3), Some(v4)) => format!("{}-{}({}-{})", v1, v2, v3, v4),
+            #[cfg(feature = "simd")]
+            (Some(v1), Some(v2), Some(v3), Some(v4)) => {
+                format!("{}-{}-simd({}-{})", v1, v2, v3, v4)
+            }
+            _ => String::new(),
+        };
+        ver
+    };
+}
 
 pub const KVSRV_API_HOST: &str = "KVSRV_API_HOST";
 pub const KVSRV_API_PORT: &str = "KVSRV_API_PORT";
