+++ metasrv/Cargo.toml
@@ -49,7 +49,6 @@ clap = { version = "3.0.0-rc.7", features = ["derive", "env"] }
 derive_more = "0.99.17"
 futures = "0.3.18"
 indexmap = "1.7.0"
-lazy_static = "1.4.0"
 maplit = "1.0.2"
 metrics = "0.17.1"
 metrics-exporter-prometheus = "0.7.0"
@@ -65,7 +64,6 @@ thiserror = "1.0.30"
 threadpool = "1.8.1"
 tokio-stream = "0.1.8"
 tonic = { version = "0.6.2", features = ["tls"] }
-
 sha2 = "0.10.0"
 uuid = { version = "0.8.2", features = ["serde", "v4"] }
 poem = { version = "1.2.4", features = ["rustls"] }
@@ -76,7 +74,3 @@ common-meta-api = { path = "../common/meta/api" }
 
 pretty_assertions = "1.0.0"
 reqwest = { version = "0.11.7", features = ["json"] }
-
-[build-dependencies]
-common-building = { path = "../common/building" }
-tonic-build = "0.6.2"
