+++ metasrv/src/configs/config.rs
@@ -16,7 +16,6 @@ use clap::Parser;
 use common_exception::ErrorCode;
 use common_meta_raft_store::config as raft_config;
 use common_meta_raft_store::config::RaftConfig;
-use lazy_static::lazy_static;
 use serde::Deserialize;
 use serde::Serialize;
 
@@ -29,26 +28,6 @@ macro_rules! load_field_from_env {
                 .parse::<$field_type>()
                 .expect(format!("cannot convert {} to {}", $env, stringify!($field_type)).as_str());
         }
-    };
-}
-
-lazy_static! {
-    pub static ref DATABEND_COMMIT_VERSION: String = {
-        let build_semver = option_env!("VERGEN_BUILD_SEMVER");
-        let git_sha = option_env!("VERGEN_GIT_SHA_SHORT");
-        let rustc_semver = option_env!("VERGEN_RUSTC_SEMVER");
-        let timestamp = option_env!("VERGEN_BUILD_TIMESTAMP");
-
-        let ver = match (build_semver, git_sha, rustc_semver, timestamp) {
-            #[cfg(not(feature = "simd"))]
-            (Some(v1), Some(v2), Some(v3), Some(v4)) => format!("{}-{}({}-{})", v1, v2, v3, v4),
-            #[cfg(feature = "simd")]
-            (Some(v1), Some(v2), Some(v3), Some(v4)) => {
-                format!("{}-{}-simd({}-{})", v1, v2, v3, v4)
-            }
-            _ => String::new(),
-        };
-        ver
     };
 }
